% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{classfd}
\alias{classfd}
\title{Performs functional discrimination of a sample of curves when a categorical response is observed (supervised classification).
A local bandwidth (i.e. local number of neighbours) is selected by a cross-validation procedure with a quadratic kernel.}
\usage{
classfd(Classes, CURVES, PRED)
}
\arguments{
\item{Classes}{vector containing the categorical responses giving the group number for each}

\item{Curve}{matrix containing the curves dataset (row by row) used for the estimating stage}

\item{Pred}{matrix containing new curves stored row by row used for computing predictions}

\item{..."}{arguments needed for the call of the function computing the semi-metric between curves}
}
\value{
A component list with:
\itemize{
\item{"Estimated.classnumber""}{Vector containing estimated class membership for each curve}
\item{"Predicted.classnumber"}{if PRED different from CURVES, this vector contains predicted class membership for each curve of PRED}
\item{"Misclas"}{misclassification rate computed from estimated values and observed values}
}
}
\description{
Performs functional discrimination of a sample of curves when a categorical response is observed (supervised classification).
A local bandwidth (i.e. local number of neighbours) is selected by a cross-validation procedure with a quadratic kernel.
}
\examples{
\dontrun{
tr <- sample(1:100, 50)
AllGwdprotein<-cbind(Gwdprotein$Group1,Gwdprotein$Group2)
train<- AllGwdprotein[,tr]
test<- AllGwdprotein[,-tr]
Classlearn <- sort(rep(1:2,25))
Classifmplsr <- classfd(Classlearn, train, test)
}
}
\references{
Ferraty, F. and Vieu, P. (2006). Nonparametric functional data analysis. Springer Series in Statistics, New York.
}

