% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{densityScores}
\alias{densityScores}
\title{densityScores Compute the Density Scores}
\usage{
densityScores(pcaobj, bivariate = FALSE)
}
\arguments{
\item{pcaobj}{PcaClass Object}

\item{bivariate}{If bivariate = TRUE compute a comtour plot base on the two Scores}
}
\value{
The component list is:
\itemize{
 \item{Estimate}{ Density Estimate}
 \item{h}{ Scalar bandwithd }
}
}
\description{
densityScores Compute the Density Scores
}
\examples{
\dontrun{
fdvector<-fdensity( fdaobjMale, pcaobj, bandwith=800)
plot(fdvector, type="p")
}
}
\references{
Delaigle, A. and Hall, P. (2010). Defining probability density for a distribution of random functions. Ann. Statist., 38, 1171-1193.
}

