% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{dpout}
\alias{dpout}
\title{dpout Graphical Bayesian selection of functional PCs.}
\usage{
dpout(lambda, plotting = TRUE, ...)
}
\arguments{
\item{lambda}{Eigenvalues of covariance or correlation matrix. (Must be in decreasing order and strictly positive -- drop the zeros)}
}
\value{
A list containg the following components:
\itemize{
\item{fhat}{Function "F-hat"}
\item{ijump}{Indices of THMAX where d-function has a step.}
\item{bstck}{ Broken stick}
}
}
\description{
dpout Graphical Bayesian selection of functional PCs.
}
\examples{
\dontrun{
Var<-varfd((Smoothfda$YSmooth))
Eg<-eigen(Var)
Dplt<- dpout(Eg$val, plotting=TRUE)
}
}
\references{
Auer, P. and Gervini, D. (2008). Choosing principal components: a new graphical method based on Bayesian model selection.  Communications in Statistics - Simulation and Computation 37 962-977.

NJ Higham (1988, Linear Algebra Appl. 103:103-118).
}

