% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{fdaclass}
\alias{fdaclass}
\title{fdaclass - Create an functional data object}
\usage{
fdaclass(mdata, argval = NULL, rangeval = NULL)
}
\arguments{
\item{mdata}{Typically a matrix of (n,m) dimension which contains a set of n curvesdiscretized in m points or argvals. Missing values are not accepted}

\item{argval}{Locations of the discritization points. Missing values are not accepted.}

\item{rangeval}{Range of discretization points, by default range(argevals). Missing values are not accepted.}
}
\value{
An object of class FDA. The component list is:
\itemize{
\item{"data"}{The data}
\item{"argvals"}{Locations of the discretization points, T1=t1, ..., Tm=tm}
\item{"rangevals"}{Range of discretization points}
}
}
\description{
fdaclass - Create an functional data object
}
\examples{
\dontrun{
fdaobjMale<-fdaclass(Gwd$Male, Gwd$Age, c(1,18))
}
}

