% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{fdensity}
\alias{fdensity}
\title{fdadensity - Compute the Functional Data Density Estimation base on the Harmonics}
\usage{
fdensity(fdaobj, pcaobj, bandwith)
}
\arguments{
\item{fdaobj}{fdClass Object}

\item{pcaobj}{PcaClass Object}

\item{bandwithd}{Bandwith for Density Estimation}
}
\value{
A vector with the estimate density values
}
\description{
fdadensity - Compute the Functional Data Density Estimation base on the Harmonics
}
\examples{
\dontrun{
fdvector<-fdensity( fdaobjMale, pcaobj, bandwith=800)
plot(fdvector, type="p")
}
}
\references{
Delaigle, A. and Hall, P. (2010). Defining probability density for a distribution of random functions. Ann. Statist., 38, 1171-1193.
}

