% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{fderiv}
\alias{fderiv}
\title{fderiv- Compute the numerical derivative for a Functional Data Object}
\usage{
fderiv(fdaobj, nderiv = 1, ...)
}
\arguments{
\item{fdaobj}{fdClass Object}

\item{nderiv}{Number of derivative to be compute}
}
\value{
A fda objects
}
\description{
fderiv- Compute the numerical derivative for a Functional Data Object
}
\examples{
\dontrun{
 FWd<-fdaclass(nonworking$data)
 Fderiv<-fderiv(FWd,1)
 matplot(Fderiv$data, type="l")
}
}
\references{
Delaigle, A. and Hall, P. (2010). Defining probability density for a distribution of random functions. Ann. Statist., 38, 1171-1193.
}

