% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{gcsvc}
\alias{gcsvc}
\title{gcsvc- Computes General Cross Validation for finding the Smoothing Parameter}
\usage{
gcsvc(fdaobj, norder, lambda, Lf, loglam)
}
\arguments{
\item{fdaobj}{An object of the FDA Class.}

\item{norder}{The order of the B-spline basis functions. The order less one is the degree of the piece-wise polynomials that make up any B-spline function.}

\item{lambda}{A nonnegative real number specifying the amount of smoothing to be applied to the estimated functional parameter.}

\item{Lf}{either a nonnegative integer or a linear differential operator object}
}
\value{
A component list with:
\itemize{
\item{"argvals "}{Range of values t_{1}, ..., t_{n}}
\item{"fdSmooth"}{Smooth Data Matrix}
\item{"df"}{Degrees of freedom}
\item{"SSE"}{Sum of the Square for Error}
\item{"PenMat"}{A roughness penalty matrix}
\item{"GCV"}{The generalized cross validation score}
}
}
\description{
gcsvc- Computes General Cross Validation for finding the Smoothing Parameter
}
\examples{
\dontrun{
norder<-6
lambda <- 0.01
Lf<-4
fdaobj<-fdaobjMale
Intv<-c(-6,0)
GCV <- gcsvc( fdaobj, norder, lambda, Lf, Intv  )
matplot(BSplinesGrowth$fdSmooth$coefs, type="l",main="Smooth Data using B-Splines",xlab="Age (years)", ylab="Heigth (cm)")
}
}
\references{
Ramsay, James O., and Silverman, Bernard W. (2006), Functional Data Analysis, 2nd ed., Springer, New York.
}

