% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{hselect}
\alias{hselect}
\title{CVSelect - Select the Cross-Validation Bandwith described in (Foster, and ) for the Median of the PSE funcion based on Functional Data}
\usage{
hselect(fdaobj, degree, interval = NULL, ...)
}
\arguments{
\item{fdaobj}{FddaObject}

\item{degree}{Degree of the local Polynomial to be used. If Degree is missing takes by default degree = 1.}

\item{Interva}{Interval to search the smoothing parameter}
}
\value{
A bandwith that minimizes the Median of the Median PSE.
}
\description{
CVSelect - Select the Cross-Validation Bandwith described in (Foster, and ) for the Median of the PSE funcion based on Functional Data
}
\examples{
\dontrun{
Int<-c(1,2)
h<- hselect(NoxNoOutliers, 1, Int)
SmoothNox<- smoothfda(fdaNox, bandwidth= h, degree=1)
matplot(SmoothNox$YSmooth, type="l",ylab="x(t)",main="Observations")
lines(SmoothNox$Mean,col="black",lwd = 4)
}
}
\references{
Peter Foster PhD Thesis. University of Manchester
}

