% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{medianfd}
\alias{medianfd}
\title{Medianfd - Computes the functional spatial median of the functional data}
\usage{
medianfd(fdaobj)
}
\arguments{
\item{time}{An object of the FDA Class.}

\item{xmat}{n object of the FDA Class.}
}
\value{
A component list with:
\itemize{
\item{"median"}{Spatial Median (1xm) vector}
\item{"weights"}{Weights (nx1) vector.}
\item{"norms"}{Distances between each curve and the median (nx1) vector. Note: The median minimizes the sum(norms)}
}
}
\description{
Medianfd - Computes the functional spatial median of the functional data
}
\examples{
\dontrun{
#Computing the Median
matplot(time,CVScores$YSmooth, type="l")
lines(time, Median$median, col="red",lwd = 4)
}
}
\references{
Gervini, D. (2008). Robust functional estimation using the spatial median and spherical principal components. Biometrika 95 587-600.
}

