% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{medianPSE}
\alias{medianPSE}
\title{CVSelect - Select the Cross-Validation Bandwith described in (Foster, and ) for the Median of the PSE funcion based on Functional Data}
\usage{
medianPSE(bandwidth, x, y, degree)
}
\arguments{
\item{bandwith}{}

\item{x}{Location of the discretization points. THis discretization points must be uniform and missing values are not accepted.}

\item{y}{Typically a matrix or data frame which contains a set of curves stored in rows. Missing values are not accepte.}

\item{degree}{Degree of the local Polynomial to be used. If Degree is missing takes by default degree = 1.}
}
\value{
A bandwith that minimizes the Median of the Median PSE for the functional data set.
}
\description{
CVSelect - Select the Cross-Validation Bandwith described in (Foster, and ) for the Median of the PSE funcion based on Functional Data
}
\examples{
\dontrun{
Mat<- fdaobjMale$data
h<- cv.select(c(0,10), 1:31,t(Mat),1)
}
}
\references{
Foster and Stehpen. PhD Thesis. Manchester University
}

