% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{outliergram}
\alias{outliergram}
\title{OutlierGram - OutlierGram}
\usage{
outliergram(fdaobj, remove = TRUE, mag = T, factorsh = 1.5,
  factormg = 1.5, p1 = 1, p2 = 0, ...)
}
\arguments{
\item{fdaobj:}{Fda Object}

\item{remove:}{Remove Outlier Option}

\item{mag:}{whether to look for magnitude outliers (mag=T) or not (mag=F).}

\item{factorsh:}{value of the factor for the detection of shape outlier.}

\item{factormg:}{value of the factor for the detection of magnitude outlier.}

\item{p1:}{at each time point, curves with values in the p1 percentile while be shifted downwards and the (mbd,mei)
will be recalculated in the new position.}

\item{p2:}{at each time point, curves with values in the p2 percentile while be shifted upwards and the (mbd,mei)
will be recalculated in the new position.}
}
\value{
The component list is:
\itemize{
 \item{ data}{Data without Outliers}
 \item{ modified.band.depth}{Modified Band depth Index}
 \item{ modified.band.depth}{Modified Band depth Index}
 \item{ modified.epigraph.index}{Modified Epigraph Index}
 \item{ dist}{ Distance }
 \item{ magnitude.outliers}{Curves that exibit a magnitude Outlier}
 \item{ shape.outliers}{Curves that exhibit a Shape Outlier }
 \item{ outliers }{Outliers Curves}
}
}
\description{
OutlierGram - OutlierGram
}
\examples{
\dontrun{
## OutlierGram
fdaobjMale<-fdaclass(Gwd$Male,Gwd$Age,c(1,18))
plot(fdaobjMale)
OG<-outliergram(fdaobjMale, remove=TRUE)
OG$shape.outliers
OG$magnitude.outliers
OG$outliers
}
}
\references{
{Shape Outlier Detection and Visualisation for Functional Data: the Outliergram. Ana Arribas-Gil and Juan Romo.
Biostatistics, (DOI) 10.1093/biostatistics/kxu006, 2014}
}

