% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{pcafd}
\alias{pcafd}
\title{pcafd - Functional Principal components analysis aims to display types of variation across a sample of
functions. Principal components analysis is an exploratory data analysis that tends to be an early part of many projects.}
\usage{
pcafd(fdaobj, nharm, Plot = FALSE)
}
\arguments{
\item{fdaobject}{A functional Data Object}

\item{nharm}{the number of harmonics or principal components to compute}
}
\value{
A list containg the following components:
\itemize{
\item{Harmonics}{ A functional data object for the harmonics or eigenfunctions.}
\item{values}{ The complete set of eigenvalues.}
\item{scores}{ A matrix of scores on the principal components or harmonics.}
\item{varprop}{A vector giving the proportion of variance explained by each eigenfunction.}
\item{cumvarprop}{A vector giving the cumulative proportion of variance explained by each eigenfunction.}
}
}
\description{
pcafd - Functional Principal components analysis aims to display types of variation across a sample of
functions. Principal components analysis is an exploratory data analysis that tends to be an early part of many projects.
}
\examples{
\dontrun{
FdaMaleHarm <- pcafd( fdaobjMale, nharm=3, Plot= TRUE)
plot(FdaMaleHarm)
}
}
\references{
Ramsay, James O., and Silverman, Bernard W. (2002), Applied Functional Data Analysis, Springer, New York.
}

