% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{pdfclasf}
\alias{pdfclasf}
\title{Performs Near Perfect Classification for functional data}
\usage{
pdfclasf(Data, test, ind0, ind1, indt)
}
\arguments{
\item{Classes}{vector containing the categorical responses giving the group number for each}

\item{"Data"}{matrix containing the curves dataset (row by row) used for the estimating stage}

\item{"test"}{matrix containing the data test.}

\item{"ind0"}{vector containing the index of the class 0}

\item{"ind1"}{vector containing the index of the class 1}

\item{"indt"}{vector containing the index of the test class}
}
\value{
A component list with:
\itemize{
\item{"Hatpsir"}{Data projection onto a space of dimension 1}
\item{"Missclasification"}{misclassification rate computed from estimated values and observed values}
\item{"MeanClass0"}{Mean for data in class 0}
\item{"MeanClass1"}{Mean for data in class 1}
}
}
\description{
Performs Near Perfect Classification for functional data
}
\examples{
\dontrun{
fd3<-pdfclasf(XXdata, XXtest, indClass0, indClass1, indtest )
plot(fd3$Hatpsir, type="l")
fd3$Missclassification
}
}
\references{
Delaigle, A. and Hall, P. (2012). Achieving near-perfect classification for functional data. J. Roy. Statist. Soc. Ser. B, 74, 267-286
}

