% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{simulatefda}
\alias{simulatefda}
\title{simulatefda - Performs simulation of Functional Data}
\usage{
simulatefda(nsamples, ndrawn, rangeval, mean, sigma)
}
\arguments{
\item{nsamples{}{Number of Sample Curves}}

\item{ndrawn{}{Number of Drawn in each sample}}

\item{rangeval{}{Range of simulation}}

\item{mean{}{Constant Mean}}

\item{sigma{}{Variance }}
}
\value{
A component list with:
\itemize{
\item{"RangeTime"}{ Simulation Range of Time }
\item{"Simulation"}{ Simulate Curves}
}
}
\description{
simulatefda - Performs simulation of Functional Data
}
\examples{
\dontrun{
simulation1<- simulatefda( nsamples=10, ndrawn=100, rangeval= c(-10,10), mean=0, sigma=1 )
simulation2<- simulatefda( nsamples=50, ndrawn=100, rangeval= c(-5,5), mean=3, sigma=1 )
n<-150
hues = seq(15, 375, length=n+1)
color<-hcl(h=hues, l=65, c=100)[1:n]
par(mfrow = c(2,1))
matplot(simulation1$rangetime,simulation1$simulation,type="l",col =color,xlab="",ylab="x(t)",main="Observations",lty=1)
matplot(simulation2$rangetime,simulation2$simulation,type="l",col =color,xlab="",ylab="x(t)",main="Observations",lty=1)
}
}
\references{
Ramsay, James O., and Silverman, Bernard W. (2006), Functional Data Analysis, 2nd ed., Springer, New York.
}

