% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{varfd}
\alias{varfd}
\title{Varfd - Computes the nearest positive definite variance - covariance matrix of functional data matrix}
\usage{
varfd(mat)
}
\arguments{
\item{mat}{Typically a matrix or data frame which contains a set of curves stored in rows. Missing values are not accepte.}
}
\value{
A list containg the following components:
\itemize{
\item{"PositDefMat"}{A Functional Variance Covariance Matrix }
}
}
\description{
Varfd - Computes the nearest positive definite variance - covariance matrix of functional data matrix
}
\examples{
\dontrun{
Mat<- fdaobjMale$data
time<- fdaobjMale$argvals
CVScores<-cv.score( 0.2, time, t(Mat))
}
}
\references{
NJ Higham (1988, Linear Algebra Appl. 103:103-118).
}

